+++ packages/sdk-router/src/index.ts
@@ -1,2 +1,3 @@
 export * from './constants'
 export * from './sdk'
+export * from './types'
