+++ packages/sdk-router/src/types/misc.ts
@@ -1,0 +1,1 @@
+export { Slippage } from '../swap'
